update home-page layout and other fixes
